+++ src/main.py
@@ -1,11 +1,12 @@
-from .agents import research_orchestrator
+from .agents.research_orchestrator import ResearchOrchestrator
 from config import GEMINI_API_KEY
 
 request_to_agent: str = "Nvidia stock price"
 def main():
     if not GEMINI_API_KEY:
-        Exception()
-    research_results: str = research_orchestrator.research_process_loop(request_to_agent, GEMINI_API_KEY)
+        raise Exception()
+    orchestrator = ResearchOrchestrator(GEMINI_API_KEY)
+    research_results: str = orchestrator.run(request_to_agent)
     print(research_results)
 
 if __name__ == "__main__":